requirements.txt
streamlit
pandas
pyodbc
faker
numpy
matplotlib
sqlalchemy
python-dotenv
openai>=1.0.0
azure-messaging-webpubsubservice
requests
pyjwt